// src/interfaces/spotify.ts
export interface SingleAlbumResponse extends AlbumObjectFull { }
// ...
interface AlbumObjectFull extends AlbumObjectSimplified {
    artists: ArtistObjectSimplified[];
    copyrights: CopyrightObject[];
    external_ids: ExternalIdObject;
    genres: string[];
    popularity: number;
    release_date: string;
    release_date_precision: string;
    tracks: PagingObject<TrackObjectSimplified>;
}

/**
 * Simplified Album Object
 * [album object (simplified)](https://developer.spotify.com/web-api/object-model/#album-object-simplified)
 */
interface AlbumObjectSimplified {
    album_type: string;
    available_markets?: string[];
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    images: ImageObject[];
    name: string;
    type: 'album';
    uri: string;
    [rest:string]:any;
}

/**
 * Full Artist Object
 * [artist object (full)](https://developer.spotify.com/web-api/object-model/)
 */
interface ArtistObjectFull extends ArtistObjectSimplified {
    followers: FollowersObject;
    genres: string[];
    images: ImageObject[];
    popularity: number;
}

/**
 * Simplified Artist Object
 * [artist object (simplified)](https://developer.spotify.com/web-api/object-model/)
 */
export interface ArtistObjectSimplified {
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    name: string;
    type: 'artist';
    uri: string;
}
// ...
interface CopyrightObject {
    text: string;
    type: 'C' | 'P';
}
// ...
interface ExternalIdObject {
    isrc?: string;
    ean?: string;
    upc?: string;
}
// ...
export interface PagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string;
    offset: number;
    previous: string;
    total: number;
}
// ...
interface TrackObjectSimplified {
    artists: ArtistObjectSimplified[];
    available_markets?: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    is_playable?: boolean;
    linked_from?: TrackLinkObject;
    name: string;
    preview_url: string;
    track_number: number;
    type: 'track';
    uri: string;
}